+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Add `withPrefixAndSuffix` param for `formatRelativeDateTime` function
+- Support short relative date time for ES locale
+- Support locale (Georgian - ka, Pashto - ps, Tagalog - tl_PH)
+
 ## 0.0.1
 
 - Initial release